HTTP Basics
This just lays down some of the groundwork for making HTTP queries.  I
have a function that will make a call to the /votes API endpoint, but
not interpret the result yet.  Just to make sure it works I threw in a
call to it in main.go that I'll remove as soon as I build an actual
HTTP handler to test with.